Move show no records button to Filter panel, and update filter logic.
